# docs/swagger.yaml
basePath: /api
definitions:
...
  internal_transport_handler.clientDTO:
    properties:
      expiry:
        type: string
      id:
        type: string
      inboundId:
        type: string
      name:
        type: string
      startAfterFirstUse:
        type: boolean
      status:
        type: string
      subscription:
        type: string
      totalQuota:
        type: integer
      usedDown:
        type: integer
      usedUp:
        type: integer
      uuid:
        type: string
    type: object
...
  internal_transport_handler.coreStatusDTO:
    properties:
      pid:
        type: integer
      running:
        type: boolean
      uptimeSeconds:
        type: integer
      version:
        type: string
    type: object
  internal_transport_handler.createClientRequest:
    properties:
      expiry:
        type: string
      inboundId:
        type: string
      name:
        type: string
      startAfterFirstUse:
        type: boolean
      totalQuota:
        type: integer
    type: object
...
  internal_transport_handler.setStatusRequest:
    properties:
      status:
        type: string
    type: object
...
  internal_transport_handler.updateClientRequest:
    properties:
      expiry:
        type: string
      inboundId:
        type: string
      name:
        type: string
      startAfterFirstUse:
        type: boolean
      status:
        type: string
      totalQuota:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: Local-first web panel for managing a sing-box process.
  title: SingGrok API
  version: 0.1.0
paths:
...
  /clients:
    get:
      parameters:
      - description: Filter by inbound ID
        in: query
        name: inboundId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/internal_transport_handler.clientDTO'
            type: array
      security:
      - BearerAuth: []
      summary: List clients
      tags:
      - clients
    post:
      consumes:
      - application/json
      parameters:
      - description: Client
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handler.createClientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_transport_handler.clientDTO'
      security:
      - BearerAuth: []
      summary: Create a client
      tags:
      - clients
  /clients/{id}:
    delete:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a client
      tags:
      - clients
    get:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.clientDTO'
      security:
      - BearerAuth: []
      summary: Get a client
      tags:
      - clients
    put:
      consumes:
      - application/json
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Patch
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handler.updateClientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.clientDTO'
      security:
      - BearerAuth: []
      summary: Update a client
      tags:
      - clients
...
  /clients/{id}/reset-traffic:
    post:
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.clientDTO'
      security:
      - BearerAuth: []
      summary: Reset a client's traffic counters
      tags:
      - clients
  /clients/{id}/status:
    post:
      consumes:
      - application/json
      parameters:
      - description: Client ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_transport_handler.setStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.clientDTO'
      security:
      - BearerAuth: []
      summary: Set a client's status
      tags:
      - clients
  /core/config:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Preview the generated config
      tags:
      - core
  /core/reload:
    post:
      description: Renders config from the database, runs `sing-box check`, installs
        it and reloads the core. A failed check returns 400 with the validation error
        and leaves the live config untouched.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Regenerate, validate and apply the config
      tags:
      - core
  /core/restart:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Restart the core
      tags:
      - core
  /core/start:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Start the core
      tags:
      - core
  /core/status:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_transport_handler.coreStatusDTO'
      security:
      - BearerAuth: []
      summary: Core process status
      tags:
      - core
  /core/stop:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stop the core
      tags:
      - core
...
securityDefinitions:
  BearerAuth:
    description: Enter "Bearer " followed by the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"